requirements.txt
pyrogram~=2.0.106
python-dotenv~=1.0.1
black

aiogram~=3.13.1